refactor: rename loginactivity to setupactivity
Since it no longer just hosts login, change to instead be SetupActivity to reflect its purpose
better
