+++ app/src/androidTest/java/com/chesire/malime/flow/login/SyncingTests.kt
@@ -15,6 +15,7 @@ import com.chesire.malime.core.api.UserApi
 import com.chesire.malime.core.flags.Service
 import com.chesire.malime.core.models.ImageModel
 import com.chesire.malime.core.models.UserModel
+import com.chesire.malime.flow.SetupActivity
 import com.chesire.malime.flow.OverviewActivity
 import com.chesire.malime.helpers.createSeriesModel
 import com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed
@@ -31,7 +32,7 @@ import javax.inject.Inject
 @RunWith(AndroidJUnit4::class)
 class SyncingTests {
     @get:Rule
-    val loginActivity = ActivityTestRule(LoginActivity::class.java, false, false)
+    val loginActivity = ActivityTestRule(SetupActivity::class.java, false, false)
 
     @Inject
     lateinit var auth: AuthApi
