+++ app/src/main/AndroidManifest.xml
@@ -50,7 +50,7 @@
         </activity>
 
         <activity android:name=".flow.OverviewActivity" />
-        <activity android:name=".flow.login.LoginActivity" />
+        <activity android:name=".flow.SetupActivity" />
     </application>
 
 </manifest>
